requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Computer Vision and AI
opencv-python==4.8.1.78
ultralytics==8.0.196
torch==2.1.0
torchvision==0.16.0

# AI and Natural Language Processing
openai>=1.20.0,<2.0.0

# Streamlit and WebRTC
streamlit>=1.28.0
streamlit-webrtc>=0.47.0
av>=10.0.0
pillow>=9.0.0

# HTTP Client for API integration
requests>=2.31.0

# Utilities
python-dotenv==1.0.0
numpy<2.0.0

# Optional: For advanced logging and monitoring
python-multipart==0.0.6